ViewPostCard のリファクタリングを実施
