+++ src/app/profile/[identifier]/page.tsx
@@ -45,6 +45,7 @@ import Link from "next/link"
 import { ListFooterNoContent } from "@/app/_components/ListFooterNoContent"
 import { useCurrentMenuType } from "@/app/_atoms/headerMenu"
 import { ViewPostCard, ViewPostCardProps } from "@/app/_components/ViewPostCard"
+import { processPostBodyText } from "@/app/_lib/post/processPostBodyText"
 
 export default function Root() {
     const [, setCurrentMenuType] = useCurrentMenuType()
@@ -261,6 +262,10 @@ export default function Root() {
                     const postProps: ViewPostCardProps = {
                         isTop: false,
                         isMobile,
+                        bodyText: processPostBodyText(
+                            nextQueryParams,
+                            post.post
+                        ),
                         postJson: post.post,
                         now,
                         nextQueryParams,
@@ -282,6 +287,7 @@ export default function Root() {
                     isTop: false,
                     isSkeleton: true,
                     isMobile,
+                    bodyText: undefined,
                     now,
                     nextQueryParams,
                     t,
